\name{complement}
\alias{complement}
\title{Graph complement
}
\description{
Complement of the graph.
}
\usage{
complement(g,n=Order(g))
}
\arguments{
  \item{g}{
a graph.
}
  \item{n}{
the order of the graph.
}
}
\details{
A graph with an edge if and only if the edge is not in \code{g}.
}
\value{
a graph.
}
\references{
Douglas B. West, Introduction to Graph Theory, Prentice-Hall, 2001.
}
\author{
David Marchette, dmarchette@gmail.com
}

\keyword{ math }
