\name{rtree}
\alias{rtree}
\title{
Random tree.
}
\description{
Generate a random tree.
}
\usage{
rtree(n)
}
\arguments{
  \item{n}{
order of the graph.
}
}
\details{
This generates a random tree.
}
\value{
a graph.
}
\author{
David Marchette, dmarchette@gmail.com
}

\seealso{
\code{\link{graphs}}, \code{\link{rg}}
}
\keyword{ math }
