\name{rtree}
\alias{rtree}
\title{
Random tree.
}
\description{
Generate a random tree.
}
\usage{
rtree(n)
}
\arguments{
  \item{n}{
order of the graph.
}
}
\details{
This generates a random tree.
}
\value{
a graph.
}
\author{
David Marchette, david.marchette@navy.mil
}

\seealso{
\code{\link{rchordal}}
}
\keyword{ graphs }
